A......... [ZBXNEXT-1161] fixed parent and dependency updating via the service.update method
git-svn-id: svn://svn.zabbix.com/branches/dev@26901 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-1161/frontends/php/api/classes/CService.php
@@ -251,21 +251,31 @@ class CService extends CZBXAPI {
 		$dependencies = array();
 		$parentDependencies = array();
 		$serviceTimes = array();
+		$deleteParentsForServiceIds = array();
+		$deleteDependenciesForServiceIds = array();
 		foreach ($services as $service) {
-			if (!empty($service['dependencies'])) {
-				foreach ($service['dependencies'] as $dependency) {
-					$dependency['serviceid'] = $service['serviceid'];
-					$dependencies[] = $dependency;
+			if (isset($service['dependencies'])) {
+				$deleteDependenciesForServiceIds[] = $service['serviceid'];
+
+				if ($service['dependencies']) {
+					foreach ($service['dependencies'] as $dependency) {
+						$dependency['serviceid'] = $service['serviceid'];
+						$dependencies[] = $dependency;
+					}
 				}
 			}
 
 			// update parent
-			if (!empty($service['parentid'])) {
-				$parentDependencies[] = array(
-					'serviceid' => $service['parentid'],
-					'dependsOnServiceid' => $service['serviceid'],
-					'soft' => 0
-				);
+			if (isset($service['parentid'])) {
+				$deleteParentsForServiceIds[] = $service['serviceid'];
+
+				if ($service['parentid']) {
+					$parentDependencies[] = array(
+						'serviceid' => $service['parentid'],
+						'dependsOnServiceid' => $service['serviceid'],
+						'soft' => 0
+					);
+				}
 			}
 
 			// save service times
@@ -278,9 +288,11 @@ class CService extends CZBXAPI {
 		}
 
 		// replace dependencies
-		$this->deleteParentDependencies(zbx_objectValues($services, 'serviceid'));
-		if ($dependencies) {
-			$this->deleteDependencies(array_unique(zbx_objectValues($dependencies, 'serviceid')));
+		if ($deleteParentsForServiceIds) {
+			$this->deleteParentDependencies(zbx_objectValues($services, 'serviceid'));
+		}
+		if ($deleteDependenciesForServiceIds) {
+			$this->deleteDependencies(array_unique($deleteDependenciesForServiceIds));
 		}
 		if ($parentDependencies || $dependencies) {
 			$this->addDependencies(array_merge($parentDependencies, $dependencies));
